lab3/skeleton/backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-stubs==5.0.2
django-stubs-ext==5.0.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.0
drf-spectacular==0.27.2
idna==3.8
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mypy==1.10.1
mypy-extensions==1.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
sqlparse==0.5.1
tblib==3.0.0
types-PyYAML==6.0.12.20240808
types-requests==2.32.0.20240907
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
